Cleanup warnings

+++ examples/thesis/edge_seg.cpp
@@ -2,6 +2,7 @@
 #include <edge/nfa.hpp>
 #include <edge/nms.hpp>
 #include <edge/threshold.hpp>
+#include <edge/draw.hpp>
 #include <geometry/draw.hpp>
 #include <imgproc/derivative_gradient.hpp>
 #include <eval/eval_app.hpp>
@@ -40,7 +41,7 @@ class EdgeSegApp : public EvalApp {
 
   using ConsoleAppInterface::run;
 
-  cv::RNG rng{time(0)};
+  cv::RNG rng{static_cast<uint64>(time(0))};
 
   void defineArgs() {
     ConsoleApp::defineArgs();
